Solution to Day 06, Part 2

+++ src/main/java/de/co/ret/day06/RaceRecord.java
@@ -1,22 +1,22 @@
 package de.co.ret.day06;
 
 import java.util.List;
-import java.util.stream.IntStream;
+import java.util.stream.LongStream;
 
-public record RaceRecord(int totalTimeMillis, int bestOfDistanceMillimeter) {
-    public List<Integer> calculateAllRaceOptions() {
+public record RaceRecord(int totalTimeMillis, long bestOfDistanceMillimeter) {
+    public List<Long> calculateAllRaceOptions() {
         return getRaceOptions()
                 .boxed().toList();
     }
 
-    public List<Integer> calculateWinningDistances() {
+    public List<Long> calculateWinningDistances() {
         return getRaceOptions()
                 .filter(distanceOption -> distanceOption > bestOfDistanceMillimeter)
                 .boxed().toList();
     }
 
-    private IntStream getRaceOptions() {
-        return IntStream.range(0, totalTimeMillis + 1)
+    private LongStream getRaceOptions() {
+        return LongStream.range(0, totalTimeMillis + 1)
                 .map(time -> time * (totalTimeMillis - time));
     }
 }
